+++ app/frontend/components/domains/requirements-library/requirement-field-edit/editable-group.tsx
@@ -6,10 +6,10 @@ import { EEnergyStepCodeDependencyRequirementCode } from "../../../../types/enum
 import { isStepCodePackageFileRequirementCode } from "../../../../utils/utility-functions"
 import { EditableHelperText, TEditableHelperTextProps } from "./editable-helper-text"
 import { EditableLabel, TEditableLabelProps } from "./editable-label"
-import { IsElectiveCheckbox, TIsElectiveCheckboxProps } from "./is-elective-checkbox"
+import { IsElectiveCheckbox } from "./is-elective-checkbox"
 import { IsMultipleFilesCheckbox } from "./is-multiple-files-checkbox"
-import { IsOptionalCheckbox, TIsOptionalCheckboxProps } from "./is-optional-checkbox"
-import { TIsMultipleFilesCheckboxProps } from "./types"
+import { IsOptionalCheckbox } from "./is-optional-checkbox"
+import { TIsElectiveCheckboxProps, TIsMultipleFilesCheckboxProps, TIsOptionalCheckboxProps } from "./types"
 
 export type TEditableGroupProps<TFieldValues extends FieldValues> = {
   editableLabelProps?: TEditableLabelProps<TFieldValues>
@@ -50,11 +50,11 @@ export function EditableGroup<TFieldValues>({
           <EditableHelperText {...editableHelperTextProps} />
         </Stack>
       )}
-      <IsOptionalCheckbox isDisabled={isEditLimited} {...isOptionalCheckboxProps} />
-      <IsElectiveCheckbox isDisabled={isEditLimited} mt={2} {...isElectiveCheckboxProps} />
       {isMultipleFilesCheckboxProps && (
         <IsMultipleFilesCheckbox isDisabled={isEditLimited} mt={2} {...isMultipleFilesCheckboxProps} />
       )}
+      <IsOptionalCheckbox isDisabled={isEditLimited} mt={2} {...isOptionalCheckboxProps} />
+      <IsElectiveCheckbox isDisabled={isEditLimited} mt={2} {...isElectiveCheckboxProps} />
     </Stack>
   )
 }
