+++ app/frontend/components/domains/requirements-library/requirement-field-edit/is-optional-checkbox.tsx
@@ -1,11 +1,8 @@
-import { Checkbox, CheckboxProps } from "@chakra-ui/react"
+import { Checkbox } from "@chakra-ui/react"
 import React from "react"
 import { FieldValues, useController } from "react-hook-form"
 import { useTranslation } from "react-i18next"
-import { IControlProps } from "./types"
-
-export type TIsOptionalCheckboxProps<TFieldValues extends FieldValues> = IControlProps<TFieldValues> &
-  Partial<CheckboxProps>
+import { TIsOptionalCheckboxProps } from "./types"
 
 export function IsOptionalCheckbox<TFieldValues extends FieldValues>({
   controlProps,
